+++ firebase-firestore/src/main/java/com/androidhuman/rxfirebase2/firestore/SetDocumentObserver.java
@@ -1,0 +1,52 @@
+package com.androidhuman.rxfirebase2.firestore;
+
+import com.google.android.gms.tasks.Task;
+import com.google.firebase.firestore.DocumentReference;
+
+import com.androidhuman.rxfirebase2.core.OnCompleteDisposable;
+
+import android.support.annotation.NonNull;
+
+import io.reactivex.Completable;
+import io.reactivex.CompletableObserver;
+
+final class SetDocumentObserver<T> extends Completable {
+
+    private final DocumentReference instance;
+
+    private final T value;
+
+    SetDocumentObserver(DocumentReference instance, T value) {
+        this.instance = instance;
+        this.value = value;
+    }
+
+    @Override
+    protected void subscribeActual(CompletableObserver observer) {
+        Listener listener = new Listener(observer);
+        observer.onSubscribe(listener);
+
+        instance.set(value)
+                .addOnCompleteListener(listener);
+    }
+
+    private static final class Listener extends OnCompleteDisposable<Void> {
+
+        private final CompletableObserver observer;
+
+        Listener(@NonNull CompletableObserver observer) {
+            this.observer = observer;
+        }
+
+        @Override
+        public void onComplete(@NonNull Task task) {
+            if (!isDisposed()) {
+                if (!task.isSuccessful()) {
+                    observer.onError(task.getException());
+                } else {
+                    observer.onComplete();
+                }
+            }
+        }
+    }
+}