+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Change Log
 
+## firestore-17.0.2.0, storage-16.0.1.0 *(2018/xx/xx)*
+
+Built and tested with:
+- Firebase Android SDK
+  - `firebase-firestore:17.0.1`
+  - `firebase-storage:16.0.1`
+- Kotlin 1.2.41
+- RxJava2 2.1.8
+
 ## core-16.0.1.0, auth-16.0.2.0 *(2018/06/21)*
 
 Built and tested with: